IMPORTANT - Firefox Patch Keep
Since this bug only shows on firefox browsers and appears to be a browser-specific bug, temporary patch is to disable the dark mode toggle element entirely for users who use firefox (Desktop and Mobile).

Patched bug is that when user clicks toggle to turn on dark mode, page jumps to bottom scroll position. Header then is no longer a fixed element. User refreshes page and bug is fixed but dark mode not set (dark-mode cookie was never set), or user can re-toggle the slider to fix the bug, again with no dark mode set.

Patch is commented in header.php, and consists of simply adding inline CSS to say display: none; on the dark toggle element if the PHP script detects the user is using firefox. To be un-patched if I stumble upon a fix or if the firefox bug is patched.