Articles Logic Fix
Articles page was still checking for the "post" GET parameter being set in order to display a specific article. Now, it correctly checks for the "id" GET parameter being set.

(This commit was required because the "post" parameter is now optional, and is only used as a URL slug for users. It is not required for the database query.)