Dark Mode Slider Inactive on 1st Click - Fix
Previously, when a user would visit the website for the first time, they would not have a dark mode cookie set. If they clicked the slider, it would check if the cookie was set to off, and since it was not off, it would hit the else block and simply set the cookie to off. Then when the JavaScript would update the color scheme afterwards, it would see that the cookie was set to off, and so it would not update to dark for the user on that first click. Now, updateCookie() checks if getCookie() has returned the empty string corresponding to no cookie value, and will set the cookie to on, as intended.