Remove --main-spacing Increase, Tweak Projects and Featured Projects
Remove Increasing the --main-spacing CSS variable in the Desktop Breakpoint because I was always having to revert it back for elements' desktop layout.

Standardized the projects page and featured projects section on the homepage as "article-card" elements. Projects page is within an "article-list" element, whereas featured projects are within an "article-carousel" element.